Move init_kerberos_ticket() to common.utils
